add custom validation

+++ api/Controllers/CommentController.cs
@@ -14,22 +14,34 @@ namespace api.Controllers
     public class CommentController : ControllerBase
     {
         private readonly ICommentRepository _commentRepo;
+        // we want a class var for the stock repo as well so that we can check if the stock exists
+        private readonly IStockRepository _stockRepo;
 
-        public CommentController(ICommentRepository commentRepo)
+        public CommentController(ICommentRepository commentRepo, IStockRepository stockRepo)
         {
             _commentRepo = commentRepo;
+            _stockRepo = stockRepo;
         }
 
         [HttpGet]
         public async Task<IActionResult> GetAll(){
+            // modelstate ensures that that the data model (dtos in our case) passes all validation
+            if(!ModelState.IsValid){
+                return BadRequest(ModelState);
+            }
             var comments = await _commentRepo.GetAllAsync();
             // DTO (Data Transfer Object) is just a way to map the data
             var commentDto = comments.Select(c => c.ToCommentDto());
 
             return Ok(commentDto);
         }
-        [HttpGet("{id}")]
+        // using the semicolon to specify a type is just adding a simple type constraint
+        [HttpGet("{id:int}")]
         public async Task<IActionResult> GetById([FromRoute] int id){
+            // modelstate ensures that that the data model (dtos in our case) passes all validation
+            if(!ModelState.IsValid){
+                return BadRequest(ModelState);
+            }
            var comment = await _commentRepo.GetByIdAsync(id);
             if(comment == null){
                 return NotFound();
@@ -37,9 +49,17 @@ namespace api.Controllers
             return Ok(comment.ToCommentDto());
         }
 
-        [HttpPost]
-        public async Task<IActionResult> Create([FromBody] CreateCommentRequestDto commentDto){
-            var commentModel = commentDto.ToCommentFromCreateDTO();
+        [HttpPost("{stockId:int}")]
+        public async Task<IActionResult> Create([FromRoute] int stockId, [FromBody] CreateCommentRequestDto commentDto){
+            // modelstate ensures that that the data model (dtos in our case) passes all validation
+            if(!ModelState.IsValid){
+                return BadRequest(ModelState);
+            }
+            // check if stock exists
+            if(!await _stockRepo.stockExists(stockId)){
+                return BadRequest("stock does not exist");
+            }
+            var commentModel = commentDto.ToCommentFromCreateDto(stockId);
             commentModel = await _commentRepo.CreateAsync(commentModel);
             // This created at action is going to perform getbyid, with the param provided in the second arg, and tostockdto is going to be the final action
             return CreatedAtAction(nameof(GetById), new {id = commentModel.Id}, commentModel.ToCommentDto());
@@ -47,8 +67,12 @@ namespace api.Controllers
         }
 
         [HttpPut]
-        [Route("{id}")]
+        [Route("{id:int}")]
         public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateCommentRequestDto updateDto){
+            // modelstate ensures that that the data model (dtos in our case) passes all validation
+            if(!ModelState.IsValid){
+                return BadRequest(ModelState);
+            }
             var commentModel = await _commentRepo.UpdateAsync(id, updateDto);
             if(commentModel == null){
                 return NotFound();
@@ -57,8 +81,12 @@ namespace api.Controllers
         }
 
         [HttpDelete]
-        [Route("{id}")]
+        [Route("{id:int}")]
         public async Task<IActionResult> Delete([FromRoute] int id){
+            // modelstate ensures that that the data model (dtos in our case) passes all validation
+            if(!ModelState.IsValid){
+                return BadRequest(ModelState);
+            }
             var commentModel = await _commentRepo.DeleteAsync(id);
             if(commentModel == null){
                 return NotFound();
