+++ api/Controllers/StockController.cs
@@ -28,6 +28,10 @@ namespace api.Controllers
         [HttpGet]
         // iaction result is defined on the entity framework to hold the results of an http request
         public async Task<IActionResult> GetAll(){
+            // modelstate ensures that that the data model (dtos in our case) passes all validation
+            if(!ModelState.IsValid){
+                return BadRequest(ModelState);
+            }
             // LINQ queries are deferred by nature meaning that the results are not returned until they're actually needed by the application
             // using ToList forces them to be retrieved at this specific point when we're calling this
             // so essentially we're bypassing the defered and saying that we want this right now
@@ -38,9 +42,13 @@ namespace api.Controllers
             return Ok(stocks);
         }
 
-        [HttpGet("{id}")]
+        [HttpGet("{id:int}")]
         // returning one record with the FromRoute
         public async Task<IActionResult> GetById([FromRoute] int id){
+            // modelstate ensures that that the data model (dtos in our case) passes all validation
+            if(!ModelState.IsValid){
+                return BadRequest(ModelState);
+            }
            var stock = await _stockRepo.GetByIdAsync(id);
             if(stock == null){
                 return NotFound();
@@ -51,6 +59,10 @@ namespace api.Controllers
         [HttpPost]
         // fromBody is an annotation saying that whatever is passed in the body of the request will exist in that param
         public async Task<IActionResult> Create([FromBody] CreateStockRequestDto stockDto){
+            // modelstate ensures that that the data model (dtos in our case) passes all validation
+            if(!ModelState.IsValid){
+                return BadRequest(ModelState);
+            }
             var stockModel = stockDto.ToStockFromCreateDTO();
             stockModel = await _stockRepo.CreateAsync(stockModel);
             // This created at action is going to perform getbyid, with the param provided in the second arg, and tostockdto is going to be the final action
@@ -59,9 +71,13 @@ namespace api.Controllers
         }
 
         [HttpPut]
-        [Route("{id}")]
+        [Route("{id:int}")]
         public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateStockRequestDto updateDto){
-            var stockModel = await _stockRepo.UpdateAsync(id, updateDto);
+           // modelstate ensures that that the data model (dtos in our case) passes all validation
+            if(!ModelState.IsValid){
+                return BadRequest(ModelState);
+            }
+             var stockModel = await _stockRepo.UpdateAsync(id, updateDto);
             if(stockModel == null){
                 return NotFound();
             }
@@ -69,8 +85,12 @@ namespace api.Controllers
         }
 
         [HttpDelete]
-        [Route("{id}")]
+        [Route("{id:int}")]
         public async Task<IActionResult> Delete([FromRoute] int id){
+            // modelstate ensures that that the data model (dtos in our case) passes all validation
+            if(!ModelState.IsValid){
+                return BadRequest(ModelState);
+            }
             var stockModel = await _stockRepo.DeleteAsync(id);
             if(stockModel == null){
                 return NotFound();
